requirements.txt
paho-mqtt
streamlit
pandas
pydeck
numpy
